add presentation package

+++ domain/src/error.rs
@@ -1,6 +1,6 @@
 pub enum DomainError {
     Validation(String),
-    NotFound { entity_type: String, entity_id: i32 },
+    NotFound { entity_type: String, entity_id: i64 },
     Infrastructure(anyhow::Error),
     Unexpected(String),
 }
